+++ README.md
@@ -2,5 +2,7 @@
 2. [ array-insert-shift](./array-insert-shift/array-insert-shift.md)
 3. [ array-binary-search](./array-binary-search/array-binary-search.md)
 4. [ stack_and_queue](./stack_and_queue/stack_and_queue.md)
+5. [ linked_list_zip](./linked-list-zip/zipLists.md)
+
